+++ src/js/common/lib/JSONStorage.js
@@ -1,3 +1,25 @@
-export const get = (key) => JSON.parse(window.localStorage.getItem(key))
+class JSONStorage {
+  constructor(namespace, defaultValue = null) {
+    if (!namespace) throw new Error()
 
-export const set = (key, value) => window.localStorage.setItem(key, JSON.stringify(value))
+    this.defaultValue = defaultValue
+    this.namespace = namespace
+  }
+
+  get() {
+    const unparsedValue = window.localStorage.getItem(this.namespace)
+
+    // most likely unset
+    if (typeof unparsedValue !== 'string') {
+      return this.defaultValue
+    }
+
+    return JSON.parse(unparsedValue)
+  }
+
+  set(value) {
+    window.localStorage.setItem(this.namespace, JSON.stringify(value))
+  }
+}
+
+export default JSONStorage