+++ src/js/popup/functions/dom.js
@@ -5,8 +5,56 @@ import {
 import * as css from '../../common/lib/css'
 
 import bookmarkItemStyles from '../../../css/popup/bookmark-item.css'
+import bookmarkTreeStyles from '../../../css/popup/bookmark-tree.css'
 import editorStyles from '../../../css/popup/editor.css'
 import panelStyles from '../../../css/popup/panel.css'
+
+export function getBookmarkListEls() {
+  const getBookmarkListElsFromPanel = (...args) => {
+    const selector = args
+      .map(getClassSelector)
+      .join(' > ')
+    const els = document.querySelectorAll(selector)
+    return Array.from(els)
+  }
+
+  const bookmarkListElsInMaster = getBookmarkListElsFromPanel(
+    panelStyles.master,
+    bookmarkTreeStyles.main,
+    bookmarkTreeStyles.list
+  )
+  const bookmarkListElsInSlave = getBookmarkListElsFromPanel(
+    panelStyles.slave,
+    bookmarkTreeStyles.main,
+    bookmarkTreeStyles.list
+  )
+
+  // it should NEVER happen
+  if (
+    bookmarkListElsInMaster.length < bookmarkListElsInSlave.length ||
+    bookmarkListElsInMaster.length - bookmarkListElsInSlave.length > 1
+  ) {
+    throw new Error()
+  }
+
+  const bookmarkListEls = []
+  const totalLength = bookmarkListElsInMaster.length + bookmarkListElsInSlave.length
+  for (let i = 0; i < totalLength; i += 1) {
+    const el = i % 2 === 0 ?
+      bookmarkListElsInMaster.shift() :
+      bookmarkListElsInSlave.shift()
+    bookmarkListEls.push(el)
+  }
+
+  return bookmarkListEls
+}
+
+function getClassSelector(className) {
+  // remove class name from compose
+  className = className.split(' ')[0]
+
+  return `.${className}`
+}
 
 export function resetBodySize() {
   const bodyStyle = document.body.style
@@ -41,13 +89,6 @@ export function setPredefinedStyleSheet(options) {
     fontSize,
     setWidth
   } = options
-
-  const getClassSelector = (className) => {
-    // remove class name from compose
-    className = className.split(' ')[0]
-
-    return `.${className}`
-  }
 
   const itemHeight = getItemHeight(options)
   const itemOffsetHeight = getItemOffsetHeight(options)